requirements.txt
boto3==1.23.3
botocore==1.26.3
PyYAML==6.0
